AutoLCA/backend/requirements.txt
PyQt5==5.15.9
NodeGraphQt==0.6.44
brightway2==2.4.6
numpy
scipy
pyinstaller
jinja2
weasyprint
pandas
fastapi
uvicorn
sqlalchemy
scikit-learn
python-multipart
reportlab>=4.0.0
psycopg2-binary
python-dotenv
python-jose[cryptography]
passlib[bcrypt]
boto3
